% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoscaling.R
\name{autoscaling}
\alias{autoscaling}
\title{Auto-scaling}
\usage{
autoscaling(x)
}
\arguments{
\item{x}{m-by-n data of class \code{data.frame} or \code{matrix}. m is the number of samples (observations) and n is the number of variables.}
}
\value{
Auto-scaled data
}
\description{
Auto-scaling
}
\examples{
dat.as <- autoscaling(dat)
}

