% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baplot.R
\name{baplot}
\alias{baplot}
\title{Bland-Altman plot}
\usage{
baplot(y, yhat, sample = seq(1, length(y), by = 1), nsd)
}
\arguments{
\item{y}{actual values}

\item{yhat}{predicted values}

\item{sample}{Name of samples}

\item{nsd}{Number of standard deviations for determining the limits of agreement.}
}
\value{
Bland-Altman plot
}
\description{
Bland-Altman plot
}
\examples{
baplot(y, yhat, nsd = 3)
}

