% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdat.R
\name{extdat}
\alias{extdat}
\title{Extract the specific range of column from a data set.}
\usage{
extdat(x, start, end)
}
\arguments{
\item{x}{an object of class \code{data.frame} or \code{matrix}.}

\item{start}{name of the 1st column of new data set to be extracted.}

\item{end}{name of the last column of new data set to be extracted.}
}
\value{
a new data set of class \code{data.frame} or \code{matrix}
}
\examples{
data(peach)
x <- extdat(peach$NIR, start = 700, end = 1050)

}

