% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matsgolay.R
\name{matsgolay}
\alias{matsgolay}
\title{Savitzky-Golay smoothing, 1st/2nd derivative}
\usage{
matsgolay(x, p = 2, n = 11)
}
\arguments{
\item{x}{x m-by-n data of class \code{data.frame} or \code{matrix} to be filtered. m is the number of samples (observations) and n is the number of variables.}

\item{p}{filter order for Savitzky-Golay smoothing}

\item{n}{filter length (window size) for Savitzky-Golay smoothing (must be odd. default value is 11)}
}
\value{
an object of class \code{list} containing the smoothed, 1st derivative and 2nd derivative data
}
\description{
The function \code{matsgolay} applys Savitzky-Golay smoothing, 1st/2nd derivative to the data by using \link{\code{sgolayfilt}} in 'signal' package.
}
\examples{
x.sg <- mysgolay(x, p = 2, n = 11)
}
\seealso{
\code{\link{sgolayfilt}}
}

