% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncompopt.R
\name{ncompopt}
\alias{ncompopt}
\title{Determine the optimal number of components to include in the model}
\usage{
ncompopt(x)
}
\arguments{
\item{x}{RMSECV values calculated by the partial least squares regression}
}
\value{
the optimal number of components
}
\description{
The function \code{ncompopt} determines the optimal number of components to include the model.
}
\details{
If \code{ncomp = "auto"}, the number of components with the minimum RMSECV is selected.
If the selected n-th component with the minimum RMSECV does not improve the RMSECV by at least 2 percent, then (n-1)th compoment is selected and so forth.
}
\examples{
data(peach)
peach.pls <- plsr(Brix ~ NIR, ncomp = 15, data = peach, validation = "CV")
ncompopt(RMSEP(peach.pls)$val[2,,])

}
\seealso{
\code{\link{rowpeak}}
}

