% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plsrPlot.R
\name{plsrPlot}
\alias{plsrPlot}
\title{Drawing important plots for the partial least squares regression at once}
\usage{
plsrPlot(formula, data, testdata = NULL, ncomp = "auto", maxcomp = 10,
  output = FALSE, dir = NULL, return.stats = FALSE, ...)
}
\arguments{
\item{formula}{a model formula (see below).}

\item{data}{an optional data frame with the data to fit the model form.}

\item{testdata}{data set for prediction.}

\item{ncomp}{the number of components (latent variables) to include in the model (see below).}

\item{maxcomp}{Maximum number of components (latent variables) to be used for cross-validation.}

\item{output}{if \code{TRUE}, the results are output as PDF and CSV files.}

\item{dir}{path to the directory where the results are output (default value is \code{"PLSR_result"}).}

\item{return.stats}{if \code{TRUE}, an object of class \code{data.frame} containing the statistics of PLS regression is returned.}

\item{...}{additional arguments passed to the \code{\link{plsr}} in 'pls' package.}
}
\value{
If \code{return.stats = TRUE}, an object of class \code{data.frame} containing the statistics of PLS regression is returned.
Otherwise, an object of class \code{mvr} is returned (see \code{\link{plsr}}).
}
\description{
The function \code{plsrPlot} performs the partial least squares (PLS) regression by using the function \code{\link{plsr}} in 'pls' package and draws four kinds of plots
(cross-validation, variable importance [regression coefficient and variable importance in projection], loading, actual vs. predicted value) at once.
}
\details{
The \code{formula} argument should be a symbolic formula of the form \code{response ~ terms},
where \code{response} is the name of the response vector
and \code{terms} is the name of the predictor matrix, e.g., water ~ FTIR.
See \code{\link{plsr}} and \code{\link{lm}} for a detailed description.

If \code{ncomp = "auto"}, the optimum number of components is automatically selected (see \code{\link{ncompopt}}).
}
\examples{
result <- plsrPlot(y = dat$Brix, x = x, ncomp = "auto", output = TRUE)

result <- plsrPlot(y = dat.train$Brix, x = x.train,
                   y.test=dat.test$Brix, x.test=x.test, ncomp = 5)
}
\seealso{
\code{\link{ncompopt}}, \code{\link{plsr}}, \code{\link{lm}}
}

