% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivityRatio.R
\name{selectivityRatio}
\alias{selectivityRatio}
\title{Calculates selectivity ratio for a given regression model}
\usage{
selectivityRatio(mvr, ncomp = mvr$ncomp)
}
\arguments{
\item{mvr}{an \code{mvr} object. The fitted model built by \code{\link{plsr}} function (see \code{\link{plsr}}).}

\item{ncomp}{a positive integer. The number of components included in the model.}
}
\value{
selectivity ratio for each variable included in the model
}
\examples{
data(peach)
peach.pls <- plsr(Brix ~ NIR, ncomp = 5, data = peach, validation = "CV")
selectivityRatio(peach.pls)

}
\author{
Yasuhiro Uwadaira (supported by Vipavee Trivittayasil)
}

