% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv.R
\name{snv}
\alias{snv}
\title{Standard normal variate (SNV)}
\usage{
snv(x, range = c(1, ncol(x)))
}
\arguments{
\item{x}{\code{data.frame} or \code{matrix} with spectra in columns.}

\item{range}{range of spectra used for calculating mean value and sd value. Specify as the column numbers.}
}
\value{
spectral data applied the standard normal variate
}
\description{
The function \code{snv} applys the standard normal variate to the spectral data.
}
\examples{
dat.snv <- snv(dat)

}

