% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baplot.R
\name{baplot}
\alias{baplot}
\title{Bland-Altman plot}
\usage{
baplot(y, yhat, sample = seq(1, length(y), by = 1), nsd = 3, ...)
}
\arguments{
\item{y}{actual values}

\item{yhat}{predicted values}

\item{sample}{Name of samples}

\item{nsd}{a numeric value multiplied to the standard deviations. The product is used for determining the limits of agreement.}

\item{...}{additional arguments passed to the \code{\link{plot}}.}
}
\value{
Bland-Altman plot
}
\examples{
data(peach)
peach.pls <- plsr(Brix ~ NIR, ncomp = 5, data = peach, validation = "CV")
yhat <- peach.pls$fitted.values[, , 5]
baplot(peach$Brix, yhat, sample = rownames(peach), nsd = 3)

}

