% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ncompopt.R
\name{ncompopt}
\alias{ncompopt}
\title{Determine the optimun number of components to include in the model}
\usage{
ncompopt(x)
}
\arguments{
\item{x}{RMSECV values calculated by the partial least squares regression}
}
\value{
the optimun number of components
}
\description{
The function \code{ncompopt} determines the optimum number of components to include the model.
}
\details{
If \code{ncomp = "auto"}, the number of components with the minimum RMSECV is selected.
If there is no component with the minimum RMSECV, the maximum number of components is selected.
Then, if the last component does not improve the RMSECV by at least 2 percent, the last one excluded and so forth.
}
\examples{
ncomp <- ncompopt(x)
}
\seealso{
\code{\link{rowpeak}}
}

