% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsrauto.R
\name{plsrauto}
\alias{plsrauto}
\title{PLS regressions under different combinations of X-variable range and preprocessing method}
\usage{
plsrauto(formula, data, testdata = NULL, xrange = NULL, p = 2, n = 11,
  ...)
}
\arguments{
\item{formula}{a model formula like \code{y ~ x}. See \code{\link{plsr}} and \code{\link{lm}} for a detailed description.}

\item{testdata}{data set for prediction.}

\item{xrange}{an object of class \code{list} which contains ranges of X-variables (see below).}

\item{p}{filter order for Savitzky-Golay smoothing (default value is 2).}

\item{n}{filter length (window size) for Savitzky-Golay smoothing (must be odd. default value is 11).}

\item{...}{additional arguments passed to \code{\link{plsrPlot}} and \code{\link{plsr}} in 'pls' package.}
}
\value{
an object of class \code{data.frame} containing the statistics of PLS regressions under the different combinations of X-variable range and preprocessing method is returned.
}
\description{
The function \code{plsrauto} performs the partial least squares (PLS) regressions under different combinations of X-variable range and preprocessing method automatically.
}
\details{
Three steps of preprocessing are automatically applied to the X-variable data set.
First, standard normal variate (SNV) is applied or not.
Second, Savitzky-Golay smoothing, 1st derivative or 2nd derivative is applied or not, respectively.
Finally, auto-scaling is applied or not.
Total 16 (2*4*2) kinds of preprocessing methods are applied and the results of PLS regressions are returned as an object of class \code{data.frame} and output as a CSV file.
If \code{output == TRUE}, PLS regression result per combination of a X-variable range and a preprocessing method is output as PDF and CSV files in one directory.
}
\examples{
# import CSV file
datAll <- read.csv(file = "./peachNIR.csv", row.names = 1, check.names = F)

datAll[1:5, 1:7]
Brix firmness       700       702       704       706       708
ak01 12.4     1.37 0.1264722 0.1206582 0.1162688 0.1129761 0.1105212
ak02 13.1     0.83 0.0894380 0.0878140 0.0866534 0.0858628 0.0853562
ak03 12.7     1.28 0.0887944 0.0860848 0.0840838 0.0826256 0.0815916
ak04 10.2     1.50 0.1637433 0.1554227 0.1490126 0.1440929 0.1403139
ak05 11.4     1.00 0.0983719 0.0939621 0.0906616 0.0882200 0.0864351

# extract X-variables
x <- extdat(datAll, start = 700, end = 2498)

# merge Y-variables and X-variables
dat <- data.frame(datAll[, 1:2], NIR = I(x))

# divide the data set into training and test data set
datTrain <- dat[1:50, ]
datTest  <- dat[51:74, ]

result.all <- plsrauto(Brix ~ NIR, data = datTrain, testdata = datTest,
                       xrange = list(c(700, 1098), c(1100, 2498)))

}
\seealso{
\code{\link{plsrPlot}}, \code{\link{plsr}}, \code{\link{lm}}
}

