% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsropt.R
\docType{package}
\name{plsropt}
\alias{plsropt}
\alias{plsropt-package}
\title{plsropt: A package for optimizing spectral preprocessing method for PLS regression}
\description{
plsropt package can be used to optimize spectral preprocessing method for partial least squares regression.
}

