% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowpeak.R
\name{rowpeak}
\alias{rowpeak}
\title{Number of rows with high and low peak value}
\usage{
rowpeak(x)
}
\arguments{
\item{x}{an object of class \code{vector}}
}
\value{
an object of class \code{list} containing the number of rows which have high and low peak values
}
\description{
Number of rows with high and low peak value
}
\examples{
rowpeak(x)
}

