% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowpeak.R
\name{rowpeak}
\alias{rowpeak}
\title{Number of rows with high and low peak value}
\usage{
rowpeak(x)
}
\arguments{
\item{x}{an object of class \code{vector}}
}
\value{
an object of class \code{list} containing the number of rows which have high and low peak values
}
\examples{
data(peach)
peach.pls <- plsr(Brix ~ NIR, ncomp = 15, data = peach, validation = "CV")
plot(RMSEP(peach.pls))
peak <- rowpeak(RMSEP(peach.pls)$val[2,,])
peak$low
RMSEP(peach.pls)$comps[peak$low]

}

