% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIP.R
\name{VIP}
\alias{VIP}
\title{Variable importance in projection for PLS regression}
\usage{
VIP(object)
}
\arguments{
\item{object}{a model object}
}
\value{
VIP values
}
\examples{
data(yarn)
yarn.pls <- plsr(density ~ NIR, 6, data = yarn, method = "oscorespls", validation = "CV")
VIP(yarn.pls)

}
\author{
Bjørn-Helge Mevik (bhx6@mevik.net)
}

