% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VIP.R
\name{VIP}
\alias{VIP}
\title{Variable importance in projection for PLS regression}
\usage{
VIP(object)
}
\arguments{
\item{object}{a model object}
}
\value{
VIP values
}
\description{
Variable importance in projection for PLS regression
}
\examples{
yarn.pls <- plsr(density ~ NIR, 6, data = yarn, validation = "CV")
VIP(yarn.pls)
}

